LCD Nokia 5110 shows Status, IP, GATEWAY and Channels Configuration.
There are a bug which is not fixed yet,
when the TIVA read the address of "width" of a PPSdivider all the
registers of the TINYFPGA are set to the value of the "width" and
it is not possible overwrite the registers with new values.

Thats why this version not read the "width" address of the PPSDivier.
The rest of registers works well. Except "stop_addr" and "start_addr"
of PPSdivider. read these registers has the same effect as read the "width"
register.

Problably it is a problem in the hardware architecture of the TINYFPGA.